requirements.txt
torch
torchvision
tqdm
Pillow
scipy
matplotlib